Fix modifiers sticking when grabbing input
Currently libweston first "runs" bindings and then calls grab callbacks,
so when we run them in the grab, we get bindings sent twice. That's why
we use an idle callback so that we don't send any bindings from grabs until
the grab_start is done.

Expo: we must make sure not to focus any view while we still haven't
finished the zoom out animation, so we must call set_focus(NULL) when
setting workspace and focusing the proper view in finalize_and_exit()