render-manager: implement scaled rendering in workspace_stream
When using workspace_streams, the resulting texture is often rendered
later again scaled. As up to now streams were drawn in full resolution,
this often results in much more pixels being rendered than is actually
necessary. To resolve this, workspace_streams now have scale_x and
scale_y, indicating how much the respective dimensions are scaled. These
can be dynamically changed, however they work only for scale values less
than one(scaling down).

In order to reduce unnecessary operations, the scaled rendering occurs
in the same texture as before(which has dimensions of the screen). The
actual image is in the lower-left corner of the texture. Also, to be
more efficient, the image won't be redrawn even when another scale is
specified in workspace_stream_update(), so users of this functionality
should check stream->scale_x/y for the actual scale of the image.

For more details on usage, see plugins/single_plugins/expo.cpp