gesture-recognizer: properly track which fingers have been sent
It is also important to distinguish between fingers sent to grab and
fingers sent to clients. We must be careful when tracking these in
register_touch(), because their state might change when starting a
grab/gesture. We must also send_up all fingers which are sent to clients
when starting a grab.