shell: implement shm-surface backend to work without cairo-gl
On many distros the default cairo(and even cairo-gl) packages don't
support the cairo-gl backend. That's why we use the shm backend instead.

To accomodate the change the code paths have been split into two files -
egl-backend.cpp and shm-backend.cpp. At compilation, exactly one of the
two is built and thus provides the functions for creating and
manipulating windows. Also, we have to re-create the windows on output
resize, because the shm backend doesn't allow window resize.
Lastly, the HAS_CAIRO_GL_H is added in config.h, so that apps know if
they can use it