proper handling of fullscreen/override-redirected views
When dealing wiht fullscreen views two things must be considered:
1. Panel layer should be "lowered", actually its position is unset, thus
non-autohiding panels won't obscure fullscreen views
2. When there is a single non-obscured fs view, etc. libweston uses
an optimization for direct scanout from the view's data, thus saving
power and improving performance by bypassing the gl-renderer as a whole.
However, this breaks all kinds of animation that Wayfire uses. As a
result, when we try to "start" an animation(usually we activate the
        plugin), we unfullscreen all fullscreened views to procede