view: add ping and ping-timeout support
These mechanisms are available for xdg-shell and Xwayland.

The idea is that plugins send a ping request to the client, and it
should respond within 10 seconds. If it does, nothing happens,
otherwise, the plugin(s) can guess that the client is unresponsive and
thus present the user with an option to kill that client.

Fixes #140