xwayland: Fix positioning of views so they don't change workspaces
When xwayland views make a configure request to move outside the workspace
that contains its center point, constrain the view to the workspace. This
fixes clients like gimp, that try to configure themselves to another
output, resulting in wayfire placing it on a different workspace. It also
fixes the case where an xwayland client is started and then workspace
switch happens; if the client configures itself, it would move to the
current workspace instead of staying where it started. This closes #768.