preserve-output plugin
This plugin fixes #898

After or during DPMS, some monitors/GPUs disconnect the outputs. This
can result in all of the views being on one output after DPMS.

This plugin works around this problem. When an output is destroyed, all
of the views on the output at the time have information stored on them.
This info includes:
1. Output (identified by make, model and serial)
2. Geometry
3. Minimised, fullscreen, tiled info
4. Z-order
5. Whether or not it was focused

When the output reconnects, all views that were on this output are moved
to it, and all state is restored. This store is deleted from the view if
the view is explicitly moved or resized by the user.