cursor: Use reference counter to ensure cursor remains hidden
The situation before was only two functions, hide_cursor and set_cursor. There were
no checks so anything could hide or show the cursor at any time. This meant that one
part of the code could hide the cursor and another might make a call to show it,
in direct succession. Here we introduce a new function unhide_cursor and a variable
for reference counting. Hide and unhide increment and decrement the reference counter
respectively. The set_cursor function is a noop if the hide reference counter is not 0.