cube: Pin pointer so it does not move while rotating
When using cube with plugins like follow-focus output, grabbing the cube and
moving the pointer to another output resulted in the cube not reotating until
moving back to the grabbed output. This patch makes cube hold the pointer in
place while rotating, so it cannot move to any other output. We also get rid
of the logic for storing saved cursor position.