xdg-shell: keep track of last serial
As discussed in #923, there are certain races when a view is resized
multiple times, for example in the following case:

1. Client has size 100x100
2. We send a configure for 50x50. Set pending compositor size to 50x50.
3. Clients commits before ACKing 50x50, reset pending compositor size to 100x100.
4. Plugin tries to resize to 100x100 again, but this is the same as
   compositor pending size (per 3.) so this fails
5. Client commits again and ACKs our configure request and has size 50x50.

In the end, the client has the wrong size. The solution is to keep track
of which the last configure sent by Wayfire is, and on commit, to check
if the client has caught up with us already.