image-loader: Add support for cubemap target
Make the image loader support loading cubemap textures automatically.
The cubemap texture size dimensions must match after the width is
divided by 4 and the height is divided by 3. The file extension must
be png or jpg. The resolution must not exceed that which is supported
by your gpu. Check the log for errors if a cubemap does not load.