scale: optionally show the titles of views as an overlay
Title is attached to scale's transformer as an overlay and is displayed on top of the view. In case of a view with dialogs, only the parent's title is displayed. If any of the dialogs would overlap (and thus occlude) the title, it is rendered on top of the topmost such dialog. Padding is used if necessary.

Titles can be shown for all views or for only the view that is under the mouse.