build: Add -fno-gnu-unique compiler option
dlclose POSIX docs say "Although a dlclose() operation is not required
to remove structures from an address space, neither is an implementation
prohibited from doing so."[1]. The fno-gnu-unique option makes it so the
plugin is unloaded when dlclose is called[2]. This makes wayfire use less
resident memory when unloading plugins. It also allows developers to make
changes to wayfire plugins without restarting wayfire.

1. The Open Group Base Specifications Issue 6.
http://pubs.opengroup.org/onlinepubs/007904975/functions/dlclose.html
2. GCC Code Gen Options.
https://gcc.gnu.org/onlinedocs/gcc/Code-Gen-Options.html